tilemap entity moved to perna-lib/entities
and.. I discovered a bug in Rotor, I decided to add position component
to tilemap, however I wasn't planning to already adapt the system to
this change.

However, on rotor/storages.nelua:67

```lua
local gen = self.generations[gen_idx.index]
```

so, it tries to get a invalid index, so it crashes the game.

I need to investigate why this happened, but get free time is so
difficult :(